requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
insightface>=0.7.3
onnxruntime-gpu>=1.16.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
click>=8.1.0
sqlalchemy>=2.0.0
pillow>=10.0.0
mtcnn>=0.1.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pydantic>=2.0.0